further ClipboardEditor and hooks refactoring (e.g. to get rid of some HMR warnings)
