+++ projects/bp-gallery/src/components/App.tsx
@@ -4,9 +4,9 @@ import 'react-perfect-scrollbar/dist/css/styles.css';
 import { renderRoutes } from 'react-router-config';
 import { buildHttpLink, mergeByRef, mergeByRefWrappedInData } from '../helpers/app-helpers';
 import './App.scss';
-import { ClipboardEditorProvider } from './common/clipboard/ClipboardEditorContext';
 import AlertProvider from './provider/AlertProvider';
 import AuthProvider from './provider/AuthProvider';
+import ClipboardEditorProvider from './provider/ClipboardEditorProvider';
 import DialogProvider from './provider/DialogProvider';
 import StorageProvider from './provider/StorageProvider';
 import routes from './routes';
