+++ projects/bp-gallery/src/components/provider/ShowStatsProvider.tsx
@@ -4,6 +4,8 @@ export const ShowStatsContext = createContext(false);
 
 export const ShowStatsProvider = ShowStatsContext.Provider;
 
-export const ShowStats = ({ children }: PropsWithChildren<{}>) => (
+const ShowStats = ({ children }: PropsWithChildren<{}>) => (
   <ShowStatsContext.Provider value={true}>{children}</ShowStatsContext.Provider>
 );
+
+export default ShowStats;
