+++ projects/bp-gallery/src/components/views/picture/sidebar/picture-info/LinkedInfoField.tsx
@@ -5,14 +5,13 @@ import { useCallback, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useGetMultiplePictureInfoLazyQuery } from '../../../../../graphql/APIConnector';
 import { useSimplifiedQueryResponseData } from '../../../../../graphql/queryUtils';
+import { useClipboard, useSetClipboardEditorButtons } from '../../../../../hooks/context-hooks';
 import { FlatPicture } from '../../../../../types/additionalFlatTypes';
 import CheckboxButton from '../../../../common/CheckboxButton';
-import { useSetClipboardEditorButtons } from '../../../../common/clipboard/ClipboardEditorContext';
 import { HelpTooltip } from '../../../../common/HelpTooltip';
 import PictureScrollGrid from '../../../../common/picture-gallery/PictureScrollGrid';
 import ScrollContainer from '../../../../common/ScrollContainer';
 import { AuthRole, useAuth } from '../../../../provider/AuthProvider';
-import { useClipboard } from '../../../../provider/context-hooks';
 import { DialogPreset, useDialog } from '../../../../provider/DialogProvider';
 import { ShowStatsProvider } from '../../../../provider/ShowStatsProvider';
 import './LinkedInfoField.scss';
